Move routes into their own files
Should make it easier to navigate
